move user columns and resourceArea config to userViews
