+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Adhocracy.ts
@@ -87,7 +87,8 @@ export var init = (config : AdhConfig.IService, meta_api) => {
         AdhResourceArea.moduleName,
         AdhProposal.moduleName,
         AdhSticky.moduleName,
-        AdhTracking.moduleName
+        AdhTracking.moduleName,
+        AdhUserViews
     ];
 
     if (config.cachebust) {
